Move internal audio source files to internal subdir
