+++ src/audio/internal/audioplayer.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "iaudioplayer.h"
+#include "../iaudioplayer.h"
 
 struct ma_decoder;
 struct ma_sound;
