+++ src/audio/internal/audioplayerstub.h
